dApp: use incognito mode for Cypress
Using the incognito mode makes sure that the whole environment is 100%
clean. This solves the issue that old test runs affect new ones. The
manual removal of all possible data in the `beforeAll` and `beforeAll`
hooks were not reliably working and where a additional work to maintain.

A big part of the diff within this commits comes from some code
restructuring in the plugin file. It attempts to make the core of the
plugin easier to read. After all it only got the following additions:
  - force the usage of Chromium based browser engines
  - add the incognito parameter for the browser