dApp: serve dApp manually for e2e tests
So far we were using a vue-cli plugin which was taking care of this for
us. But as we now want to have tests for the offline functionality of
the dApp, we need to manage the server manually. This allows us to serve
the dApp initially and then stop it after the dApp cached everything. If
the end-to-end tests do still pass, the service worker works correctly
(can be verified by shutting down the server too early). The actual time
when the caching is done must be estimated as there is no way to
determine that reliably. Also the super slow Cypress network connection
must be taken into account here. After all the time value must be small
enough to still test some offline functionality. This is guaranteed the
case when the tests do a backup and reload the dApp.

This orchestration process required its own script as it did not fit
nicely just in the `package.json`. Having it as its own script gives us
much more freedom (e.g. trap hooks), the ability to group code into
named functions and also add comments with important details. Especially
for such a case where we have to manage multiple background processes.