dApp: handle deployment info parsing errors
If the build dApp got configured to use some "special" deployment info
files and it fails to fetch or parse them, this results into a plain
error message. In fact it get caught by the same try-and-catch block as
the SDK initialization and we can map it to a specific, more user
friendly, message. As the causes for such errors can vary a lot, it
it is more efficient to wrap them into a new custom error class. This
preserves the origin error information, but also allows to an easy
mapping to a localized message.