dApp: fix stuck e2e tests when re-connecting
This issue already exists for a longer time and caused some flakiness
between multiple runs within the CI environment. As the debug/inspection
tools for Cypress within CI are basically zero, this is more a good
guess from watching videos of failed runs. The assumption is that there
is some duplicated routing/navigation in the dApp going on after doing
a backup of the state. This causes the dApp to partially reload. As
result the already running interactions by Cypress with the connection
manager can fail, depending on the timing.
This tries to mitigate the problem by having Cypress just wait for short
period before trying to interact with the connection manager.