dApp: decrease blockchain timeout for e2e tests
This timeout got introduced as some actions in the dApp can take a while due to
blockchain interactions. For example opening a channel until it is done. But the
so far used timeout value was set to 30min. The CircleCI job has a default
maximum runtime of 10min. This means if something in the tests goes wrong and
this results into Cypress waiting forever for something that will never happen
anymore, the tests do not finish. This means they don't actually fail but get
cancelled, which counts as failed by CircleCI. Though this is "correct" this
comes with the downside that if such errors occur early in the scenario, they
don't fail early. But much more importantly is that because CircleCI kills the
Cypress process, this does not produce a video or any report by Cypress. So the
developer is in fact blind and has no clue what actually happened, where, when,
...
This issue can be tackled from to sides. First it should be absolutely fine to
drop the blockchain timeout to just 3min. This is still sufficient to wait for
any interaction on the blockchain in our scenario. If this changes, the tests
start to fail. But we will be able to see where and why this happens and can act
accordingly to increase the value again.
The second change for this issue is to increase the timeout value for the job in
CircleCI. This makes sure that even in case that the Cypress scenario takes very
long due to tons of very long timeouts (within the new limit), it should still
always finish the job and produce a failure report that can be used to analyze
the issue. It must be said that this timeout relates to a job that did not
produce any output. As Cypress is very silent per default, this is an issue for
us. Though it would be still an issue, even if we would get logging, as in these
cases the scenario is stuck and just waits.