dApp: open developer console in end-to-end tests
Failing end-to-end tests of the dApp are hard to debug. The basically
only tool is the video that gets stored as artifact in CircleCI. But so
far it only shows the dApp itself and the special console of Cypress its
scenario output. For a better debugging experience it is quite important
to be able to inspect the actual browser console itself. Such includes
live logging of the SDK, shows errors of the dApp and also the service
worker.
In future it might be worth to checkout some Cypress plugins to include
the console output in a report file. None of the plugins I checked out
so far look too promising as they only focus on the special Cypress
logging (`cy.log`) only. But having the console in the video is already
very helpful as it also provides a direct correlation with events in
time.

This "feature" got realized by a simple Chromium browser launch option.
Furthermore this also sets a new directory for the browser where to
store its cache. Such includes for example the width of the console. To
prevent the browser from remembering such "settings" this is necessary.
Such things are not part of the incognito mode.