dApp: add explicit offline reload in e2e tests
The end-to-end tests must reliably verify that the dApp works offline.
So far this was only tested by stopping the server and make sure that
the scenario still succeeds. But Cypress is very slow on network traffic
due to its proxy. Thereby the initial caching of the service worker can
take very long as already described in earlier commits. From the
observation of many end-to-end test runs in the CI environment, it has
shown that it still sometimes happens that the server got stopped too
early. As we lack in a reliable mechanism to know when the application
got fully cached, it was a guessing game what timeout it sufficient.
To improve this and make sure the tests are not flaky due to timing
issues, this commit highly increases the timeout to stop the server.
This makes sure that the caching always succeeds. But to still have
a good test of the offline functionality, we now have extend the
scenario actively. Else we can't be 100% sure if the scenario will
still try to load assets it hasn't loaded before. Only then the cache
becomes relevant. Therefore the scenario will reload the whole dApp at
the very end. At this point in time the dApp must run offline. And
a full reload automatically leads to reloading the assets. If the
service worker does not work as intended, this will fail. If it
succeeds, the dApp works reliably offline.