dApp: increase limit of asset size for caching
The asset caching for the offline functionality is now also relevant for
not productive builds like the end-to-end tests. A key difference to
production builds here is that they don't have an optimized/minified
output. This means that especially the `chunk-vendors.js` asset becomes
very large. So large that in contrast to the production build, it hits
the already very hight asset size limit. As result it will not be cached
during the end-to-end test scenario and the offline functionality is not
fully working.