dApp: add structure logging for end-to-end tests
When watching a video of a failed end-to-end tests scenario, it is sometimes
hard to determine at which exact step in the flow the issue happened. That's one
of the downsides of having just one single big test case for everything (though
it is necessary). To help out, I added a plain logging instruction at the
beginning of every single helper method that is used in the scenario
flow. This uses the Cypress logging, to appear within the special logging window
of it.