dApp: increase time paddings for timing tests
We have some tests for the service worker assistant which verity that it
sends a certain message in a specific period. Testing the exact timing
of the interval is basically impossible. So we had to add some padding
to it to make it working well.
Unfortunately it has shown that in some rare cases after many many runs
within the CI workflow, that sometimes this padding is not enough. To
avoid flakiness in the end-to-end tests, I increased the time padding so
high, that is is still less than the expected number of intervals plus
one. This should be enough to make the tests even more stable even for
thousands of runs.