dApp: install serve application for e2e tests
The old solution to prevent timing issues was to have some additional
timeout after the `npx serve` command that gets forked into the
background. This was meant to prevent slow installations of the command
to make the Cypress command in the foreground process to fail. Checkout
the according recent commit for details.
The more explicit and save solution is to first install the application
within the foreground process. Then we can be sure that when the
background serve process gets started, it will serve the dApp right away
without any uncontrolled delay based on network connectivity.