dApp: force build mode for dApp in for e2e tests
Unfortunately is is basically not possible to run the Cypress end-to-end
tests of the dApp with different build modes. It must be the `e2e` mode.
Else the connection to the chain, public path and more do not work. Some
of them are solvable with a good amount of effort (I did before). But
finally this suffers from the issue that these dApp builds are meant to
be published.
Therefore we can't test the actual builds that will be published. We
must run the end-to-end tests in their specific environment and that
also requires a specific build. Though this build only slightly differs
and should have no affect on the final quality of the publish artifacts.
The overwriting build output in the CI job of the end-to-end tests just
don't get persistent to the workspace of the workflow, so it should not
affect other jobs which depend on the origin build mode output.