dApp: optimize CI workflow dependencies
There is a misconception that many jobs in a workflow like tests need to
respective component to be build first. That is wrong. Unit and integration
are meant to run without having the component actually build. The same
is true for our specific setup with the end-to-end tests. Thereby all
these jobs can be shifted on layer earlier in the workflow, reducing it
from 4 stages to 3. This should bring a big improvement in time for each
workflow.

Additionally I added some more restrictions to the publishing jobs. Like
the staging pipeline was publishing the build without waiting/verifying
(for the) end-to-end tests.

The linting jobs has to wait for the SDK to build first. This is so
because the dApp has a direct monorepo dependency to it. Linting would
fail for the dApp without the SDK build first.
There would be an opportunity to split the linting for each component
and then schedule them more effectively. In practice this would bring
not gain in time. The whole job takes around 1min and 15s. The actual
linting are about 40s for all components. If we would split this into
multiple jobs, this might lead to two jobs with 50s each. But as there
are other jobs in the same stage of the workflow that will always take
longer to execute, this has no affect at all to the overall execution
time. Thereby it would just add more complexity to the CI configuration
without much/any gain.