cli: change output directory
The raiden-ts and raiden-dapp workspaces use the naming `dist[:*]` for their output
directories. Only the raiden-cli used `build` and `bundle`. To make this more
uniform, this got changed to `dist` and `dist:bundle`.
This change has in mind some unification of all tasks in each workspace to
simplify the (CI) workflows.