sdk: introduce withdrawBusy and emit withdraw.failure early
At the first block inside "danger zone", we know we aren't going to try
to sent the tx (either withdraw or coop-settle), so we already now
withdraw/coop-settle failed here, and can emit failure early. But we
don't want to do that if the respective tx was sent before danger zone
and is pending, so withdrawBusy "locks" withdraw request to not fail
early, until it actually succeeds or fails on the respective txs.