sdk: replace semver contract check with getCode for method
`semver` did use `versions.json`, which is generated at build time and
don't know if runtime contracts being used do or don't support some
method. Now, we getCode of the contract and check the sighash is present
on it, which should allow this check to work with 0.37 and 0.39 runtime
contracts as long as it's built with the later.