sdk: add changelog entry and split withdraw into coopsettle epic
codeclimate was complaining about the side of withdraw with the added
coopsettle epics, so we split into its own epic. Also, coopSettleEpic
had too many returns due to the skipped states, so we replace it with
asserts and a swallowing catchError in the end.
No logic change here, just moving things around to satisfy codeclimate.