dApp: extend amount display with a label
This allows to specify an optional label for the amount display
component. Furthermore it can be set to be full width. This makes
basically only sense in combination with a label.

The display amount usage in the transfer component had to be adopted due
to these changes.

The unit tests of the amount display were extended, but also refactored
to align with the most recent standards.