dApp: extend address input to allow restriction
As he address input component can be configured to exclude a list of
addresses that are not allowed to be used, this ads the option to
restrict the input to a fixed list of allowed addresses. A practical use
case is for example the selection of a channel partner to make a deposit
to.

The unit tests of the amount input component are a nightmare. But that
became too big to easily improve them now.