dApp: add new action progress card component
This component is meant to be the unified effort of displaying the
progress of an action to the user. Actions are understood as time
consuming processes where the user has to wait for it to complete.
Examples are opening a channel or withdrawing from the UDC.
So far this is implemented everywhere multiple times over and over again
with different feature sets and slightly different visuals. This
component should unify them and replace them on a long term.

The action progress card is just for the visualization, not the action
itself. The progress consists of a overall state and the state of
multiple steps that are within this action. The overall state can be in
progress, completed or failed (with a respective error). Each step can
be pending, active, completed or failed. Steps have a title and
a description with user instructions (e.g. to confirm a blockchain
transaction). An example for steps are the opening of a channel with
a final deposit, where it firsts opens the channel, then does a token
transfer and then deposit to the channel. This component does include no
own logic to interpret data on a higher level or controls how the steps
are changing their state.

The usage of this component will follow soon in the next commits.