dApp: add new channel action form component
This component is meant to be used for handling the parameters of
a channel related action (e.g. opening a channel). It consists of the
three values that relate to channel actions which are the token
(address), the partner (address) and the token amount (e.g. to deposit).
Each value can be either just be displayed or is allowed to be edited by
the user. Values can be also completely hidden if they are clear for the
context (e.g. acting on a selected channel). Inputs can be restricted
according to special scenarios (e.g. exclude channel partners when
opening a new channel).

The form component finally provides a single component which is
responsible to fetch token information and monitor them. Before (or
actually still is yet) this was done everywhere we needed it. As this is
an asynchronous process, this was causing a lot of issues in some
components. Now they must only handle the addresses as strings and the
form will do the work.

The form will be bound to a specific action to execute. It also allows
to customize some labeling for the specific action and some other visual
adoptions. Of course it only allows to execute the action if all form
values are valid. This does not apply for fixed values that are not
editable. They are provided by the parent component and expected to be
correct. This might be spot to improve at a later point (especially if
the values come from URL parameters).

The component emits events according to the outcome of the executed
action.

Note that for the moment the token value of the form is not editable and always
fixed. This is because we don't have yet a token input component. Only the big
component for the select token route. This might be added at a later point.