dApp: add first action for channel deposit
The action progress card unifies the display of an action to the user.
The now newly added action mixin is the abstract class for all action
implementations. This means these components are not actually displaying
anything, but control the action (steps) and use the progress card to
visualize it. The first here added action which is using this setup is
the channel deposit action. It is a very simple one and consists of
a single step only.

The mixin controls the higher level action state, and the orchestration
of the control flow including the state reset and all computed
properties. The parts which are individual by the actions like the
steps are then customized by the components using the mixin.

The mixin includes a minimal visual part which allows to either use the
plain action progress card or put it into a dialog pop-up. This is
helpful as in many places we trigger actions with a button and then want
to visualize the progress independently from the existing layout (e.g.
open channel view).
In case of such a dialog it re-implements the already known feature to
automatically close the dialog (with the progress) when the action has
succeeded. The delay to do so can be customized. In case of a failure,
the user is responsible to close the dialog himself.

This also finally brings in the feature to block the user from closing
the action dialog, while the action has not yet completed

The mixin allows to display a slot component instead of the progress
while the action is not in progress. When the action gets triggered,
slot gets hidden if the progress is not showing in a dialog on top of
it.

Unfortunately Vue does not support single-file-components with
a template for mixins. This requires to use a pure render function to
provide the features of the slot and dialog handling.