dApp: add channel open action
This is the second action implementation. It re-uses the introduced
action mixin. It is more complex than the first action to do
a channel deposit as it consists of up to three steps depending on the
inputs.

This required some changes to the raiden service function to execute the
actual channel opening. The SDK accepts a event handler function as
parameter. So far this got abstracted away with a (stupid) generic
function which counts the events as steps. This got removed, allowing
the channel open action to directly update the actions progress steps up
on the emitted events by the SDK.