dApp: adapt the channel deposit dialog for action
The channel deposit dialog could be finally drastically simplified by
using the new channel deposit action and the channel action form. It
basically just puts everything into a dialog and reacts to a completed
deposit. This also reflects in the heavily simplified unit tests, as the
tests of the action and form component test all the functionality at
a single place.

This commit includes the changes to adapt the dialog usage everywhere.
This includes a lot of simplifications as the new unified component do
the heavy lifting and the using/parent components must just use them in
their templates with the correct properties.

Also the end-to-end tests got marginally adopted to some new `data-cy`
components and the handling of token amount input without a default zero
value.