dApp: add input event to channel action form
This allows the parent component of a channel action form to be update
when then input value changes. The advantages of such a single event are
that it is debounced and limited to valid inputs per default, as well as
it does not require to force the usage of synced properties and thereby
make it more optional.