sdk: UDC to withdrawToBeneficiary by default with subkey
raiden-contracts@0.39 added to UDC the ability to withdrawToBeneficiary,
removing the need to a later tx to transfer SVT from subkey back to main
account. This commit uses this by default when the subkey is used and
the contracts support it. A `options` param is added to
`Raiden.withdrawFwithdrawFromUDC` to allow passing an object containing
an explicit `subkey` boolean to force to withdraw to subkey if true (or
force to withdraw to main account if false and config.subkey is true).