dapp: hotfix for directRoute check when PFS is set
If VUE_APP_PFS is set to a free PFS, the `handleStep` check on
TransferStepsRoute doesn't wait for the `directRoute` check performed on
`create`, causing it to fail in case the direct transfer isn't possible
before having the opportunity actually perform the request. The fix is
simply to ensure the directRoute call completes before performing the
findRoutes request.