dApp: allow to make transfer path optional
There is the use case that the SDK should automatically do a transfer.
This includes the whole detection if a direct transfer is possible or
not, getting the best path finding service and the cheapest route for
a mediated transfer. This does not work if the dApps Raiden service
requires a value for the path parameter.