dApp: fix channel action form amount display
While the amount input expects to work on a plain string, the amount
display requires the BigNumber value. In the channel action form both
components got feed with the same value for the amount. This does not
work and leads to errors. But the form component is already of this type
difference and has a computed property for it. Thereby the fix was very
simple.