dApp: add error detail for target does not receive
This is a quite common error, especially for transfers to new
Raiden nodes. Even for us developers this happens from time to time in
our test setups. Then we have to waste time to find the original error
behind the else very generic error message. This shows the actual cause
and helps the user to solve the issue.