dApp: add new transfer actions
This includes actions which kinda contain of multiple actions, or rather
combine multiple actions which also exist standalone.
This allows to deposit more tokens into a channel and then do a direct
transfer through the same channel. The same works with first opening
a new channel and then a direct transfer.
This is very useful for straight forward transfers without user
interactions and a clear visual progress without interruptions.