sdk: enable stricter typescript options
`noUncheckedIndexedAccess` and `noPropertyAccessFromIndexSignature` were
added in recent TS versions (4.1 & 4.2), and allow us to be more precise
on undefined types when accessing properties from indexed types (objects
and arrays). This requires us to add strict checks/type guards to ensure
we aren't dealing with undefined values, but can save us in case we
didn't get what we thought.
This would have prevented the error fixed by #2925 , which we only
catched "in the wild" (SP).