dApp: improve error handling for quick pay route
The transfer action was handled a little special in contrast to the
other actions. This included to show the progress inline. This got
unified in combination with a better error visualization.
All actions show their progress in a dialog now. The transfer action
just has some empty translations to make it less obtrusively as intended
before already. This allows to re-use the nice looking error
visualization of the action progress card in the dialog. It provides
more information and looks better. Before this was hidden by a second
error dialog on top of it.

This change did also allow to get rid of the extra functionality to hide
and display again the action message (which was only for the inline
progress of the transfer action).

Though there is still an error dialog remaining to handle all errors
which are not caused by actions themselves. For the moment this is only
the case for invalid query parameters.

Moreover does the changing of query parameters during an active action
not cancel this action anymore. Furthermore do action errors take
precedence.

Finally the closing of the action progress dialog will also cause the
direction. This solves a missing feature where the user was stuck for
failed actions, waiting for ever to get redirected. Second this allows
the user to get quicker redirected in case of a success by closing the
dialog manually. This became possible by a recent extension of the
action mixin.