sdk: refactor: avoid race on channelDeposit
channelDeposit.request's payload can receive either `deposit` (previous)
or `totalDeposit`; in case of former, the `totalDeposit` is calculated
at request's handling time; once totalDeposit is defined, it'll be used
on contract's call, independent of external deposits happening in
parallel. To avoid races, concatMap will hold the lock until the deposit
is confirmed. `totalDeposit` then is used in channelOpen.request with
deposit, to avoid races when depositing.