sdk: allow parallel deposits per tokenNetwork if allowance is big
we were very conservative when serializing deposits per tokenNetwork, as
the allowance can conflict and waste gas, but the downside is that
deposits in different channels on the same network needs to happen
serially. We improve it here by allowing serializing per channel instead
of per tokenNetwork when `config.minimumAllowance` is the default of
MaxUint256 (so one deposit doesn't incur the risk of affecting
allowance of another parallel), making them parallel and faster to
confirm.
Also applies the same serialization to channelOpen, which were
previously unprotected and allowed any number of parallel open requests.