dApp: add custom query parameter parsing
In 'hash' mode, the router expects the URL search (query) string to be within
the hash of the URL (e.g. `http://test.tld/#/route?foo=bar`). This does not
align with the W3C standard. There everything after a not encoded '#' belongs to
the hash. The search string gets placed before the hash. As a result, anyone who
creates a link for the dApp probably ends up with
`https://test.tld/?foo=bar#route`. Unfortunately this does not get properly
parsed by the routers default `parseQuery` function.
As a solution this custom function parses both search/query strings and merges
both. Hash query parameter take precedence in case both are defined and the same
key exists twice.

This issue was discover during the testing with the PayWall prototype.