sdk: broaden transfer's options.paths to InputPaths
paths now can be a `InputPaths`, a new codec which allows a broader set
of options: actual route can come from either `route` or `path` members,
and `estimated_fee` can be received and is converted to `fee` by adding
the fee margins according to config. The `...rest` of members are passed
through. These changes allow it to be used both to decode PFS's result
as well as to receive CLI's `paths` option.