dApp: split transfer action into two
Due to recent decisions, the user must always be able to see the full
costs of his transactions before confirming them. This means that
the fully automated transfer by the SDK can not be used anymore as it
might end up in a mediated transfer with additional hidden costs.
To solve this issue the dApp must always decide to do a direct or
mediated transfer. Therefore the former transfer action got replaced by
two new actions: direct transfer and mediated transfer.
The direct transfer will also construct the to use direct route itself.
This is necessary to ensure it will be a direct transfer. Furthermore it
is fine to put this into the action itself, as it will not cause any
hidden costs. This is different for the mediated transfer. Here the user
must decide himself to first query a route (paying the pathfinding
service) and then use this route with the additional fees. Therefore the
mediated transfer action is expecting the route as parameter.