dApp: rework additional action component props
The action components, based on the action mixin are quiet some special
components. But some of them are slightly different than the others.
These are those which include (also) a transfer. So far action are quite
tightly coupled to the usage with action forms in may cases. Though this
might become improved in future, at the moment there is the need for
some actions to take more options to perform their action then the
action form provides (e.g. payment identifier). These have been provided
as special properties which are only defined on a couple of those
actions. This lead to different interfaces of the actions, which is not
nice. Especially where actions are used interchangeable like for the
quick pay feature it is necessary to have a unified interface.

So far this situation was accepted, because all actions that were
different where different the same. And on the quick pay route only
those special actions where used. But with the recent addition of direct
and mediated transfer actions, this has changed. The mediated transfer
now also needs to be provided with a route. This also causes issues in
the quick pay route.

To solve it, a new property has been added to the base action mixin.
This piece of data works the same as the generic `options` argument to
the `handleAction` function of the actions. It is an object that can be
specified individually by each implementing action. The same goes now
for the `fixedRunOptions`. They are supposed to be provided as
an optional component property and can be used as any other data in the
respective actions implementation. Those actions who don't need any
additional options just ignore them.

After all there are some downsides to this approach: type safety. Having
this generic new properties does not allow for any static type checks.
The action will basically fail (or not) when the respective action logic
can not handle the undefined parameters which are expected. But it could
also lead to unintended behavior if they can but work differently then.
After all typical JavaScript issues.
We accept these issues for the moment. The actions were not safe
beforehand as well as this is just the same approach as we took for the
direct/"flexible" action options. There are different approaches to
tackle this issue to detect miss-usage early. Though probably not on
a static code analysis base. Unfortunately none of these approaches is
really the best and it requires more thoughts how to solve it
completely with something that works for both option datasets.