dApp: add slot feature to amount display
This new features allows to provide a custom slot element to the display
amount. The optional slot will be displayed instead of the actual amount.
The intention behind this feature is to provide a very generic approach
to make the component more adaptive for special cases. An example would
be the case where the amount is not yet known and needs to be fetched
asynchronously. Then the slot can be filled with a spinner component. If
the fetching has been completed, the spinner gets removed (conditionally
rendering) and the amount gets displayed as usual. This also allows to
use the amount display component is cases where the amount might be
undefined in the beginning, but still having this component contributing
to the layout and maybe have a label to it.
The slot approach keeps the component very simple without much impact
and much customization for special scenarios only. So it remains a very
basic component that can be used everywhere.

Furthermore there were small improvements to the code itself. In the
components implementation/template as well as the tests (make them more
explicit).

Finally it is now possible to define an empty label on purpose. Doing so
allows to maintain the layout. This makes for example sense when there
are multiple display components together or the label might change other
time. If the label is completely missing (undefined) the pure amount
will be displayed as always.