dApp: adapt quick pay feature for transparency
Due to recent decisions, the user must always be able to see the full
costs of his transactions before confirming them. This means that
the fully automated transfer by the SDK can not be used anymore as it
might end up in a mediated transfer with additional hidden costs.
To solve this issue the dApp must always decide to do a direct or
mediated transfer.
As a result of this, the quick pay route must manually fetch the available
pathfinding services and select the cheapest one. In case a direct transfer is
not possible but a mediated one, the user will get displayed more payment
information. This includes a new section for the mediation. Here the user sees
the price of the cheapest service once it is known. Then the user can decide
himself to spend the money to fetch a route.
Once the possible routes are fetched, the cheapest one gets selected and the
user can see the mediation fees for this route. The user can now see the
complete of costs the payment will charge him.

While the user has not yet asked to query the pathfinding service for a route,
the quick pay view will show him the alternative payment solution. This will
either be a channel deposit and direct transfer or a channel open and direct
transfer. Once he has received a valid route, this will switch to a mediated
transfer. So far the user can not decide against the fetched route and go back
to the alternative action. As workaround he would need to reload the page and go
straight for it.

The whole section of mediation stuff includes some nice features like visual
progress states, and the detecting if a pathfinding service is too expensive
with visual effects and disabling.

The logic to handle failed transfer got removed as it is not necessary anymore.
It was only used to check that a mediated transfer was not possible and then
decide on an alternative action. This is now build in from the very beginning
and the user has the choice between both.

I'm happy about the user flow we have designed here. But I'm not too happy with
the code. The quick pay route has become way too big and complex. Though it is
quite hard to split it in a helpful way. Somehow it would be nice to outsource
the whole payment information stuff with the new mediation interactive logic.
But it is to much tied with all the other stuff that it produces more overhead
than actual gain.
Furthermore have the latest updates shown that the combination of actions with
the channel action form is not flexible enough. While a nice feature, it must
become improved with a more generic form approach.