sdk: refactor: replace `using` with `latestEpic`
we used rxjs's `using` operator to ensure `getLatest$` subscription to
`deps.latest$` was the first performed (to receive actions/state changes
notifications before actual epics) and the last one disposed. With the
epics subscription order, this can actually become almost a _normal_
epic, with the only special case that this must be the first subscribed
epic, and the last unsubscribed/shut down, which removes the explicit
subject subscription as well.