sdk: refactor: get rid of subkey options in some public methods
Some (but not all) methods of the Raiden public class exposed an
`options.subkey` parameter, which allowed to trigger these specific
operations to be performed by the subkey account (even though they'd be
sent by the main account by default), or vice versa (i.e. opting-in to
the main account tx when config.subkey=true would cause the subkey to be
used by default). As this duplicates functionality of config.subkey, and
both are very seldomly used, we remove the options.subkey per tx to
simplify the code. Users can always set config.subkey only for a tx then
reset it afterwards as needed, so functionality is kept.