sdk: refactor: rename rxjs custom utils mergeWith -> withMergeFrom
rxjs@7 renamed  `merge` operator to `mergeWith`, which merges the
content of observable arguments with input (like `merge` function, but
as inline operator), and this conflicts with the name we've choosen for
our custom operator (no functionality yet provided ootb).
The new name `withMergeFrom`, is also more descriptive, as it acts like
`withLatestFrom` operator, but without losing values from input before
arguments emitted (which the later does) and also emitting for each
inner emition as well.