sdk: refactor: move gasPrice logic to `transact` and drop epics
blockGasPriceEpic and actions were needed only because we had calling
transactions spread across different parts of code. Now that we
centralize tx calls in `transact`, we can move gasPrice[Factor] logic to
it. This also optimizes not calling `getFeeData` on every block, and
instead just when transactioning, with a local cache to ensure
concurrent calls are deduplicated.