dapp: add exception to no connected token
There is a special routing guard that checks if the user has any
connected token. If not, it will automatically forces him to route to
the screen where he can connect. There are a few exceptions to this like
the account settings.
This does also prevent a new user to use the quick pay feature right
away. But as the quick pay works also for new users, this must be a new
exception to the guard.