dapp: disable service worker for staging
The staging version that gets deployed publicly should be be always the
most recent build. There is no version file for it. So if running the
service worker, this will prevent the user from getting the most recent
version as it will have its cache and never updates, as the version file
on the server remains the same until the next actual release.