dapp: add new warning message to quick pay route
When the pathfinding servie is too expensive, the price was shown in red
and the button to fetch a route was disabled. This leaves the user with
no information why. To improve this, instead of the button a warning
message gets displayed. This bascially makes the former 'no routes'
approach more generic an allows to use it for other messages as well.
The price will not be shown in red color anymore as this would be too
much red then.